Upload temporary file - Version 1 beta release.
We wanted to get code out to you as soon as possible, however we're still working on getting it all ready to upload to GitHub. We're making sure that we've make our code as compatible as possible. So for right now, this code is ready to use, but we're still working on a couple of things.